+++ packages/fuels-core/Cargo.toml
@@ -13,7 +13,6 @@ description = "Fuel Rust SDK core."
 bech32 = "0.9.0"
 chrono = "0.4.2"
 uint = { version = "0.9.5",  default-features = false }
-impl-serde = { version = "0.4.0", default-features = false }
 fuel-abi-types = { workspace = true }
 fuel-asm = { workspace = true }
 fuel-core = { workspace = true, default-features = false, optional = true }
