+++ packages/fuels-core/src/types/core.rs
@@ -1,11 +1,13 @@
 pub use bits::*;
 pub use bytes::*;
-pub use native::*;
+pub use identity::*;
 pub use raw_slice::*;
 pub use sized_ascii_string::*;
+pub use u256::*;
 
 mod bits;
 mod bytes;
-mod native;
+mod identity;
 mod raw_slice;
 mod sized_ascii_string;
+mod u256;
